Moving logic to reporioty
