+++ lib/services/orderService.ts
@@ -4,7 +4,12 @@ import { OrderRepository } from '../repositories/orderRepository';
 export class OrderService {
     constructor(private repo: OrderRepository) {}
 
-    async placeOrder(customerEmail: string, ShippingAddress: string, OrderDetails: any[]): Promise<Order> {
-        return await this.repo.placeOrder(customerEmail, ShippingAddress, OrderDetails);
+    async placeOrder(orderId: string, orderDate: string, customerEmail: string, ShippingAddress: string, OrderDetails: any[]): Promise<Order> {
+        try {
+            return await this.repo.placeOrder(orderId, orderDate,customerEmail, ShippingAddress, OrderDetails);
+        } catch (error) {
+            console.error('Error in OrderService.placeOrder:', error);
+            throw error;
+        }
     }
 }
